feat: AccountService 리팩터링

+++ src/main/java/com/uos/all4runner/domain/dto/request/AccountRequest.java
@@ -18,7 +18,7 @@ public interface AccountRequest {
 		@NotBlank String addressGu,
 		@NotBlank String addressDong,
 		@Min(value = 0) @Max(value = 20) Double avgspeed,
-		@Min(value = 0) @Max(value = 200) Integer weight
+		@Min(value = 0) @Max(value = 200) Double weight
 		){
 	}
 }
